fix: heredoc delimiter

+++ src/parser/infile.c
@@ -68,30 +68,46 @@ int	simple_infile(char *file_name, t_cmd_info *cmd)
 	return (0);
 }
 
+static char	*get_hd_name(t_cmd_info *cmd)
+{
+	char	*hd_name;
+	char	*hd_id;
+
+	hd_id = ft_itoa(cmd->hd_count++);
+	if (!hd_id)
+		return (NULL);
+	hd_name = ft_strjoin("here_doc_", hd_id);
+	if (!hd_name)
+		return (free(hd_id), NULL);
+	free(hd_id);
+	return (hd_name);
+}
+
 int	here_doc(char *delimiter, t_cmd_info *cmd)
 {
 	char	*next_line;
 	int		fd;
 	char	*hd_name;
 
-	hd_name = ft_strjoin("here_doc_", ft_itoa(cmd->hd_count++));
+	hd_name = get_hd_name(cmd);
 	fd = open(hd_name, O_RDWR | O_CREAT | O_TRUNC | O_APPEND, 0777);
 	if (fd < 0)
-		return (set_err(cmd, ERR_NO_FILE, hd_name));
+		return (free(hd_name), set_err(cmd, ERR_NO_FILE, hd_name));
 	while (1)
 	{
 		next_line = readline(">");
-		if (ft_strncmp(delimiter, next_line, ft_strlen(delimiter)) == 0)
+		if (ft_strlen(delimiter) == ft_strlen(next_line)
+			&& ft_strncmp(delimiter, next_line, ft_strlen(delimiter)) == 0)
 		{
 			free(next_line);
 			close(fd);
 			cmd->infile = open(hd_name, O_RDONLY);
 			unlink(hd_name);
+			free(hd_name); 
 			return (0);
 		}
-		write(fd, next_line, ft_strlen(next_line));
-		write(fd, "\n", 1);
+		ft_fprintf(fd, "%s\n", next_line);
 		free(next_line);
 	}
-	return (0);
+	return (free(hd_name), 0);
 }
